This fixes the bug that stops Project.js code from running at all
The bug is about the double functionality Techie paradigm implemented recently to reduce repetition in code. This allows you to both specify a selector string and a function to run at the matched set upon document ready. This is very handy solution when you want you whole context to base on a particular selection say pt("body"). So with Techie("body", function(){}), this object automatically becomes the matched set